+++ Firestore/core/CMakeLists.txt
@@ -22,6 +22,7 @@ add_subdirectory(src/firebase/firestore/immutable)
 add_subdirectory(src/firebase/firestore/local)
 add_subdirectory(src/firebase/firestore/model)
 add_subdirectory(src/firebase/firestore/nanopb)
+add_subdirectory(src/firebase/firestore/objc)
 add_subdirectory(src/firebase/firestore/remote)
 add_subdirectory(src/firebase/firestore/util)
 
@@ -32,6 +33,7 @@ add_subdirectory(test/firebase/firestore/core)
 add_subdirectory(test/firebase/firestore/immutable)
 add_subdirectory(test/firebase/firestore/local)
 add_subdirectory(test/firebase/firestore/model)
+add_subdirectory(test/firebase/firestore/objc)
 add_subdirectory(test/firebase/firestore/nanopb)
 add_subdirectory(test/firebase/firestore/remote)
 add_subdirectory(test/firebase/firestore/util)
