+++ Firestore/core/src/firebase/firestore/objc/CMakeLists.txt
@@ -1,0 +1,27 @@
+# Copyright 2019 Google
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# firebase_firestore_util is the interface of this module. The rest of the
+# libraries in here are an implementation detail of making this a
+# mutli-platform build.
+
+cc_library(
+  firebase_firestore_objc
+  SOURCES
+    objc_compatibility.h
+    objc_type_traits.h
+    empty.mm  # cmake fails to link without at least one non-header
+  DEPENDS
+    absl_meta
+)
