+++ Firestore/core/src/firebase/firestore/util/CMakeLists.txt
@@ -250,7 +250,6 @@ cc_library(
     error_apple.mm
     hashing.h
     iterator_adaptors.h
-    objc_compatibility.h
     ordered_code.cc
     ordered_code.h
     range.h
